backend/requirements.txt
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cryptography==45.0.7
Flask==3.1.2
google-ai-generativelanguage==0.4.0
google-api-core==2.25.1
google-auth==2.40.3
google-generativeai==0.3.0
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.62.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
proto-plus==1.26.1
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pypdfium2==4.30.0
python-dotenv==1.0.0
requests==2.32.5
rsa==4.9.1
tqdm==4.67.1
urllib3==2.5.0
Werkzeug==3.1.3
